Change loading page and errors to routing pages
